enhance command palette ui ux (#3985)
* enhance command palette ui ux

* Improvise command palette UI/UX and fix workspace file search

* test(web): cover customize-store membersTab reset behaviour

Satisfies the ships-with-tests gate for the command-palette refactor by
locking the new membersTab default/reset/deep-link contract.


---------
